revisando controllers

+++ app/Http/Controllers/AnsweredTestController.php
@@ -7,14 +7,15 @@ use App\Http\Requests\AnsTestRequest;
 
 class AnsweredTestController extends Controller
 {
-    /*public function index()
+    public function index()
     {
-        $ans_Test = AnsweredTest::all();
-        return $ans_Test;
-    }*/
+
+        return redirect('/', 403);
+    }
 
     public function store(AnsTestRequest $request)
     {
+        $this->authorize('create');
         $validated = $request->validated();
         $id = $validated["school_member_id"];
         $answered = AnsweredTest::create($validated);
@@ -23,11 +24,13 @@ class AnsweredTestController extends Controller
 
     public function show(AnsweredTest $answeredTest)
     {
+        $this->authorize('view', $answeredTest);
         return view('ans_test.show', compact('answeredTest'));
     }
 
     public function update(AnsTestRequest $request, AnsweredTest $answeredTest)
     {
+        $this->authorize('update', $answeredTest);
         $validated = $request->validated();
         $answeredTest->update($validated);
         return $answeredTest;
@@ -35,17 +38,19 @@ class AnsweredTestController extends Controller
 
     public function destroy(AnsweredTest $answeredTest)
     {
+        $this->authorize('delete', $answeredTest);
         $answeredTest->delete();
         return $this->index();
     }
 
     public function edit(AnsweredTest $answeredTest)
     {
+        $this->authorize('update', $answeredTest);
         return view("ans_test.edit", compact('answeredTest'));
     }
 
     public function create()
     {
-        return view("ans_test.create");
+        return redirect('/', 403);
     }
 }
